Mbox Click Tracking Script
This script tracks clicks using the mbox "mboxClickTrack". Clicks are tracked based on the conversion metrics in a designated Adobe Target campaign. Related parameters and parameter values (ex. clicked=sign-in-button) are what the Mbox communicates with.

There is a special-case conditional which looks at a specific scenario where a form checkbox is checked, then unchecked, to ensure the latter is not counted in the Target report data.